rest_api_service/requirements.txt
flask==2.2.2
flask-sqlalchemy==3.0
werkzeug==2.2.2
requests==2.31.0
psycopg2-binary==2.9.6